omni-learner/backend/requirements/development.txt
# Development Tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3
ipython==8.20.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality
black==23.12.1
flake8==7.0.0
isort==5.13.2